+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## Improvements
 - Rename VictorOps -> Splunk On-Call
+- Implement email body decoding in the "Ping Details" dialog
 
 ## Bug Fixes
 - Fix downtime summary to handle months when the check didn't exist yet (#472)